refactored code for code-review

See merge request 2ti_project_informatica_2020_2021/groep-10/leap-webapp!35

+++ backend/src/main/java/edu/ap/group10/leapwebapp/company/CompanyController.java
@@ -1,22 +1,23 @@
 package edu.ap.group10.leapwebapp.company;
 
+import java.util.Arrays;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.server.ResponseStatusException;
 
 import edu.ap.group10.leapwebapp.confirmationtoken.ConfirmationToken;
 import edu.ap.group10.leapwebapp.confirmationtoken.ConfirmationTokenService;
 import edu.ap.group10.leapwebapp.mail.Mail;
 import edu.ap.group10.leapwebapp.mail.MailService;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @RestController
 public class CompanyController {
 
@@ -31,11 +32,14 @@ public class CompanyController {
 
   @GetMapping("/companies")
   public List<Company> getAllCompanies(@RequestParam String role) {
-    if (role.equals("application admin")) {
-      return companyService.getCompanies();
-    } else {
-      throw new AccessDeniedException("Only application admins can access this");
+    try {
+      if (companyService.checkRole(role)) {
+        return companyService.getCompanies();
+      }
+    } catch (AccessDeniedException ex) {
+      log.error(ex.getMessage());
     }
+    return Arrays.asList();
   }
 
   @PostMapping("/companies")
@@ -56,7 +60,6 @@ public class CompanyController {
     String confirmationTokenString = "http://localhost:4200/company/register/?id=" + n.getId() + "&token=" + token;
 
     Mail mail = new Mail();
-    mail.setSender("leapwebapp@gmail.com");
     mail.setReceiver("standaertsander@gmail.com, stijnverhaegen@gmail.com, yasin.khawaja@student.ap.be");
     mail.setSubject("New application from: " + companyName);
     mail.setContent("Click on this link to view the request from: " + companyName + ".\n" + confirmationTokenString);
@@ -65,57 +68,53 @@ public class CompanyController {
 
   @GetMapping("/companies/{token}")
   public Company viewCompanyApplication(@PathVariable("token") String confirmationToken, @RequestParam String role) {
-    if (role.equals("application admin")) {
+    Company c = null;
+    try {
+      if (companyService.checkRole(role)) {
       ConfirmationToken token = confirmationTokenService.getConfirmationToken(confirmationToken);
-      if (token != null) {
-        Company c = companyService.findCompany(token.getCompany().getId());
-
-        if (c != null) {
-          return c;
-        } else {
-          throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Company not found");
-        }
+      c = companyService.findCompany(token.getCompany().getId());
       }
-      throw new ResponseStatusException(HttpStatus.UNAUTHORIZED,
-          "You do not have the permission to accept or refuse a company.");
-    } else {
-        throw new AccessDeniedException("Only application admins can access this");
+      return c;
+    } catch (Exception e) {
+      log.error(e.getMessage());
+      return c;
     }
   }
 
   @PostMapping("/companies/{token}/applicationStatus")
-  public void companyApplicationResult(@PathVariable String token, @RequestParam("accepted") Boolean accepted, @RequestParam String role) {
+  public void companyApplicationResult(@PathVariable String token, @RequestParam("accepted") Boolean accepted,
+      @RequestParam String role) {
+    try {
+      if (companyService.checkRole(role)) {
+        ConfirmationToken verifyToken = confirmationTokenService.getConfirmationToken(token);
 
-    if (role.equals("application admin")) {
-    ConfirmationToken verifyToken = confirmationTokenService.getConfirmationToken(token);
+        Long companyID = verifyToken.getCompany().getId();
+        Company company = companyService.findCompany(companyID);
 
-    Long companyID = verifyToken.getCompany().getId();
-    Company company = companyService.findCompany(companyID);
+        confirmationTokenService.deleteConfirmationToken(token);
 
-    confirmationTokenService.deleteConfirmationToken(token);
+        String confirmationTokenString = "";
 
-    String confirmationTokenString = "";
+        if (accepted.booleanValue()) {
+          String tokenAdmin = confirmationTokenService.addConfirmationToken(company);
+          confirmationTokenString = "http://localhost:4200/register-useradmin?token=" + tokenAdmin;
+        } else {
+          companyService.deleteCompany(companyID);
+        }
 
-    if (accepted.booleanValue()) {
-      String tokenAdmin = confirmationTokenService.addConfirmationToken(company);
-      confirmationTokenString = "http://localhost:4200/register-useradmin?token=" + tokenAdmin;
-    } else {
-      companyService.deleteCompany(companyID);
+        Mail mail = new Mail();
+        mail.setReceiver(company.getEmail());
+        mail.setSubject("Your registration has been reviewed.");
+        mail.setContent(
+            "The admin has decided to" + (accepted.booleanValue() ? " accept " : " deny ") + "your application.\n"
+                + (accepted.booleanValue()
+                    ? "You can register your user-admin by clicking on the following link: " + confirmationTokenString
+                    : "If you think this was a mistake, then send an e-mail to leapwebapp@gmail.com"));
+
+        mailService.sendMail(mail);
+      }
+    } catch (AccessDeniedException ex) {
+      log.error(ex.getMessage());
     }
-
-    Mail mail = new Mail();
-    mail.setSender("leapwebapp@gmail.com");
-    mail.setReceiver(company.getEmail());
-    mail.setSubject("Your registration has been reviewed.");
-    mail.setContent(
-        "The admin has decided to" + (accepted.booleanValue() ? " accept " : " deny ") + "your application.\n"
-            + (accepted.booleanValue()
-                ? "You can register your user-admin by clicking on the following link: " + confirmationTokenString
-                : "If you think this was a mistake, then send an e-mail to leapwebapp@gmail.com"));
-
-    mailService.sendMail(mail);
-  } else {
-    throw new AccessDeniedException("Only application admins can access this");
-  }
   }
 }
